requirements.txt
pydantic
ffmpeg-python
